requirements.txt
faiss-cpu==1.13.2
fastapi==0.128.0
langchain==0.2.17
langchain-community==0.2.16
langchain-core==0.2.43
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
pydantic==2.12.5
pypdf==6.6.0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
streamlit==1.52.2
tenacity==8.5.0
uvicorn==0.40.0